refactor: move protocol headers and command types consts to own files
